# swagger/yaml/swagger-workflow.yaml
swagger: "2.0"
info: 
   title: "API for VMS Workflow Managment"
   description: "VMS Workflow Management APIs"
   version: 1.0.0
basePath: /api/v1
paths:
...
  /flowTask/tasks:
    get:
      tags:
      - "user task managment"
      summary: "查询user task"
      description: ""
      operationId: queryTask
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: queryTaskReq
        in: body
        required: true
        schema: 
          $ref: "#/definitions/QueryTaskReqVO"
      responses:
        200:
          description: upload bpmn file  successfully
          schema:
            $ref: "#/definitions/StandardResponse"
  /flowTask/tasks/{id}:
    delete:
      tags:
      - "user task managment"
      summary: "根据id删除user task"
      description: ""
      operationId: deleteTask
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        required: true
        type: integer
      responses:
        200:
          description: delete task successfully
          schema:
            $ref: "#/definitions/StandardResponse"          
  /flowTask/claim/{id}:
    post:
      tags:
      - "user task managment"
      summary: "认领task"
      description: ""
      operationId: claimTask
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        required: true
        type: integer
      responses:
        200:
          description: complete user task successfully
          schema:
            $ref: "#/definitions/StandardResponse"          
  /flowTask/unclaim/{id}:
    post:
      tags:
      - "user task managment"
      summary: "归还之前认领的task"
      description: ""
      operationId: unclaimTask
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        required: true
        type: integer
      responses:
        200:
          description: complete user task successfully
          schema:
            $ref: "#/definitions/StandardResponse"          
  /flowTask/complete:
    post:
      tags:
      - "user task managment"
      summary: "完成task(e.g. 包含assign,group users)"
      description: ""
      operationId: completeTask
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: completeTaskReqVO
        in: body
        required: true
        schema:
          $ref: "#/definitions/CompleteTaskReqVO"
      responses:
        200:
          description: complete user task successfully
          schema:
            $ref: "#/definitions/StandardResponse"
definitions: 
...
  QueryTaskReqVO:
    properties:
      firstResult:
        type: integer
      maxResults:
        type: integer
      assignee:
        type: string
      candidateUser:
        type: string
      candidateGroup:
        type: string
...
  CompleteTaskReqVO:
     required: 
        - taskId
        - assignee
        - isReviewPass
     properties:
      taskId:
        type: string
      assignee:
        type: string
      isReviewPass:
        type: integer
      returnStart:
        type: integer
      variables:
        type: object
        additionalProperties: 
          type: object
  StandardResponse:
    required: 
       - code
       - message
       - data
    properties:
      code:
        type: integer
      message:
        type: string
      data:
        type: object
...